Updated the sql-init file

+++ Lakeshore Market/com/online/lakeshoremarket/domain/CustomerDomain.java
@@ -29,10 +29,56 @@ public class CustomerDomain {
 		return addressID;
 	}
 
-	public void deleteCustomer(int custID) {
+	public boolean deleteCustomer(int custID) {
+		boolean isCustomerDeleted = false ;
 		custDao = new CustomerDAO();
-//		custDao.deleteCustomerAddress();
-		custDao.deleteCustomer(custID);
+		isCustomerDeleted = custDao.deleteCustomer(custID);
+		
+		//This code section is commented because we don't actually have to delete the customer instead make the customer inactive
+		/*int billingAddressID = 0;
+		int shippingAddressID = 0;
+		int funcReturn = 0;
+		billingAddressID = getCustomerBillingAddress(custID);
+		shippingAddressID = getCustomerShippingAddress(custID);
+		if(0 != billingAddressID && 0 != shippingAddressID){
+			funcReturn = custDao.deleteCustomer(custID);
+			if(0 != funcReturn){
+				int billingAddressDeleted = 0;
+				int shippingAddressDeleted = 0;
+				billingAddressDeleted = custDao.deleteCustomerAddress(billingAddressID);
+				shippingAddressDeleted = custDao.deleteCustomerAddress(shippingAddressID);
+				if(0 != billingAddressDeleted && 0 != shippingAddressDeleted){
+					isCustomerDeleted = true;
+				}
+			}
+		}*/
+		return isCustomerDeleted;
+	}
+	
+	public void deleteCustomerAddress(int custID){
+		custDao = new CustomerDAO();
+		custDao.deleteCustomerAddress(custID);
+	}
+	
+	public int getCustomerBillingAddress(int custID){
+		int billingAddressID = 0;
+		custDao = new CustomerDAO();
+		billingAddressID = custDao.getCustomerBillingAddress(custID);
+		return billingAddressID;
+	}
+	
+	public int getCustomerShippingAddress(int custID){
+		int shippingAddressID = 0;
+		custDao = new CustomerDAO();
+		shippingAddressID = custDao.getCustomerShippingAddress(custID);
+		return shippingAddressID;
+	}
+
+	public boolean getStatus(int custID) {
+		boolean isCustomerActive = false;
+		custDao = new CustomerDAO();
+		isCustomerActive = custDao.getStatus(custID);
+		return isCustomerActive;
 	}
 
 }
